Servidor/GraphQL/requirements.txt
fastapi
uvicorn
strawberry-graphql